in spel-katalog, split up App impl
